requirements.txt
# Core dependencies
crewai>=0.30.11
python-dotenv==1.0.1

# Web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# LLM providers
groq>=0.5.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# Utilities
pyyaml>=6.0.0
colorama>=0.4.6
rich>=13.0.0
click>=8.0.0
werkzeug==3.0.1
requests>=2.31.0

# Development (optional)
# pytest>=8.0.0
# black>=24.0.0
# flake8>=7.0.0
# ipython>=8.0.0